+++ src/main/java/com/barion/block_variants/data/BVRecipeProvider.java
@@ -11,10 +11,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.ItemLike;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.SlabBlock;
-import net.minecraft.world.level.block.StairBlock;
-import net.minecraft.world.level.block.WallBlock;
+import net.minecraft.world.level.block.*;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.function.Consumer;
@@ -120,24 +117,24 @@ public class BVRecipeProvider extends ExtendedRecipeProvider {
         stairSlab(consumer, BVBlocks.Stripped_Crimson_Stem_Stairs.get(), BVBlocks.Stripped_Crimson_Stem_Slab.get(), Blocks.STRIPPED_CRIMSON_STEM, false);
         stairSlab(consumer, BVBlocks.Stripped_Warped_Stem_Stairs.get(), BVBlocks.Stripped_Warped_Stem_Slab.get(), Blocks.STRIPPED_WARPED_STEM, false);
 
-        recipeWoods(consumer, BVBlocks.Oak_Wood_Stairs.get(), BVBlocks.Oak_Wood_Slab.get(), BVBlocks.Oak_Wood_Wall.get(), Blocks.OAK_WOOD, Blocks.OAK_LOG);
-        recipeWoods(consumer, BVBlocks.Spruce_Wood_Stairs.get(), BVBlocks.Spruce_Wood_Slab.get(), BVBlocks.Spruce_Wood_Wall.get(), Blocks.SPRUCE_WOOD, Blocks.SPRUCE_LOG);
-        recipeWoods(consumer, BVBlocks.Birch_Wood_Stairs.get(), BVBlocks.Birch_Wood_Slab.get(), BVBlocks.Birch_Wood_Wall.get(), Blocks.BIRCH_WOOD, Blocks.BIRCH_LOG);
-        recipeWoods(consumer, BVBlocks.Jungle_Wood_Stairs.get(), BVBlocks.Jungle_Wood_Slab.get(), BVBlocks.Jungle_Wood_Wall.get(), Blocks.JUNGLE_WOOD, Blocks.JUNGLE_LOG);
-        recipeWoods(consumer, BVBlocks.Acacia_Wood_Stairs.get(), BVBlocks.Acacia_Wood_Slab.get(), BVBlocks.Acacia_Wood_Wall.get(), Blocks.ACACIA_WOOD, Blocks.ACACIA_LOG);
-        recipeWoods(consumer, BVBlocks.Dark_Oak_Wood_Stairs.get(), BVBlocks.Dark_Oak_Wood_Slab.get(), BVBlocks.Dark_Oak_Wood_Wall.get(), Blocks.DARK_OAK_WOOD, Blocks.DARK_OAK_LOG);
-        recipeWoods(consumer, BVBlocks.Mangrove_Wood_Stairs.get(), BVBlocks.Mangrove_Wood_Slab.get(), BVBlocks.Mangrove_Wood_Wall.get(), Blocks.MANGROVE_WOOD, Blocks.MANGROVE_LOG);
-        recipeWoods(consumer, BVBlocks.Crimson_Hyphae_Stairs.get(), BVBlocks.Crimson_Hyphae_Slab.get(), BVBlocks.Crimson_Hyphae_Wall.get(), Blocks.CRIMSON_HYPHAE, Blocks.CRIMSON_STEM);
-        recipeWoods(consumer, BVBlocks.Warped_Hyphae_Stairs.get(), BVBlocks.Warped_Hyphae_Slab.get(), BVBlocks.Warped_Hyphae_Wall.get(), Blocks.WARPED_HYPHAE, Blocks.WARPED_STEM);
-        recipeWoods(consumer, BVBlocks.Stripped_Oak_Wood_Stairs.get(), BVBlocks.Stripped_Oak_Wood_Slab.get(), BVBlocks.Stripped_Oak_Wood_Wall.get(), Blocks.STRIPPED_OAK_WOOD, Blocks.STRIPPED_OAK_LOG);
-        recipeWoods(consumer, BVBlocks.Stripped_Spruce_Wood_Stairs.get(), BVBlocks.Stripped_Spruce_Wood_Slab.get(), BVBlocks.Stripped_Spruce_Wood_Wall.get(), Blocks.STRIPPED_SPRUCE_WOOD, Blocks.STRIPPED_SPRUCE_LOG);
-        recipeWoods(consumer, BVBlocks.Stripped_Birch_Wood_Stairs.get(), BVBlocks.Stripped_Birch_Wood_Slab.get(), BVBlocks.Stripped_Birch_Wood_Wall.get(), Blocks.STRIPPED_BIRCH_WOOD, Blocks.STRIPPED_BIRCH_LOG);
-        recipeWoods(consumer, BVBlocks.Stripped_Jungle_Wood_Stairs.get(), BVBlocks.Stripped_Jungle_Wood_Slab.get(), BVBlocks.Stripped_Jungle_Wood_Wall.get(), Blocks.STRIPPED_JUNGLE_WOOD, Blocks.STRIPPED_JUNGLE_LOG);
-        recipeWoods(consumer, BVBlocks.Stripped_Acacia_Wood_Stairs.get(), BVBlocks.Stripped_Acacia_Wood_Slab.get(), BVBlocks.Stripped_Acacia_Wood_Wall.get(), Blocks.STRIPPED_ACACIA_WOOD, Blocks.STRIPPED_ACACIA_LOG);
-        recipeWoods(consumer, BVBlocks.Stripped_Dark_Oak_Wood_Stairs.get(), BVBlocks.Stripped_Dark_Oak_Wood_Slab.get(), BVBlocks.Stripped_Dark_Oak_Wood_Wall.get(), Blocks.STRIPPED_DARK_OAK_WOOD, Blocks.STRIPPED_DARK_OAK_LOG);
-        recipeWoods(consumer, BVBlocks.Stripped_Mangrove_Wood_Stairs.get(), BVBlocks.Stripped_Mangrove_Wood_Slab.get(), BVBlocks.Stripped_Mangrove_Wood_Wall.get(), Blocks.STRIPPED_MANGROVE_WOOD, Blocks.STRIPPED_MANGROVE_LOG);
-        recipeWoods(consumer, BVBlocks.Stripped_Crimson_Hyphae_Stairs.get(), BVBlocks.Stripped_Crimson_Hyphae_Slab.get(), BVBlocks.Stripped_Crimson_Hyphae_Wall.get(), Blocks.STRIPPED_CRIMSON_HYPHAE, Blocks.STRIPPED_CRIMSON_STEM);
-        recipeWoods(consumer, BVBlocks.Stripped_Warped_Hyphae_Stairs.get(), BVBlocks.Stripped_Warped_Hyphae_Slab.get(), BVBlocks.Stripped_Warped_Hyphae_Wall.get(), Blocks.STRIPPED_WARPED_HYPHAE, Blocks.STRIPPED_WARPED_STEM);
+        recipeWoods(consumer, BVBlocks.Oak_Wood_Stairs.get(), BVBlocks.Oak_Wood_Slab.get(), BVBlocks.Oak_Wood_Wall.get(), BVBlocks.Oak_Wood_Fence.get(), BVBlocks.Oak_Wood_FenceGate.get(), Blocks.OAK_WOOD, Blocks.OAK_LOG);
+        recipeWoods(consumer, BVBlocks.Spruce_Wood_Stairs.get(), BVBlocks.Spruce_Wood_Slab.get(), BVBlocks.Spruce_Wood_Wall.get(), BVBlocks.Spruce_Wood_Fence.get(), BVBlocks.Spruce_Wood_FenceGate.get(), Blocks.SPRUCE_WOOD, Blocks.SPRUCE_LOG);
+        recipeWoods(consumer, BVBlocks.Birch_Wood_Stairs.get(), BVBlocks.Birch_Wood_Slab.get(), BVBlocks.Birch_Wood_Wall.get(), BVBlocks.Birch_Wood_Fence.get(), BVBlocks.Birch_Wood_FenceGate.get(), Blocks.BIRCH_WOOD, Blocks.BIRCH_LOG);
+        recipeWoods(consumer, BVBlocks.Jungle_Wood_Stairs.get(), BVBlocks.Jungle_Wood_Slab.get(), BVBlocks.Jungle_Wood_Wall.get(), BVBlocks.Jungle_Wood_Fence.get(), BVBlocks.Jungle_Wood_FenceGate.get(), Blocks.JUNGLE_WOOD, Blocks.JUNGLE_LOG);
+        recipeWoods(consumer, BVBlocks.Acacia_Wood_Stairs.get(), BVBlocks.Acacia_Wood_Slab.get(), BVBlocks.Acacia_Wood_Wall.get(), BVBlocks.Acacia_Wood_Fence.get(), BVBlocks.Acacia_Wood_FenceGate.get(), Blocks.ACACIA_WOOD, Blocks.ACACIA_LOG);
+        recipeWoods(consumer, BVBlocks.Dark_Oak_Wood_Stairs.get(), BVBlocks.Dark_Oak_Wood_Slab.get(), BVBlocks.Dark_Oak_Wood_Wall.get(), BVBlocks.Dark_Oak_Wood_Fence.get(), BVBlocks.Dark_Oak_Wood_FenceGate.get(), Blocks.DARK_OAK_WOOD, Blocks.DARK_OAK_LOG);
+        recipeWoods(consumer, BVBlocks.Mangrove_Wood_Stairs.get(), BVBlocks.Mangrove_Wood_Slab.get(), BVBlocks.Mangrove_Wood_Wall.get(), BVBlocks.Mangrove_Wood_Fence.get(), BVBlocks.Mangrove_Wood_FenceGate.get(), Blocks.MANGROVE_WOOD, Blocks.MANGROVE_LOG);
+        recipeWoods(consumer, BVBlocks.Crimson_Hyphae_Stairs.get(), BVBlocks.Crimson_Hyphae_Slab.get(), BVBlocks.Crimson_Hyphae_Wall.get(), BVBlocks.Crimson_Hyphae_Fence.get(), BVBlocks.Crimson_Hyphae_FenceGate.get(), Blocks.CRIMSON_HYPHAE, Blocks.CRIMSON_STEM);
+        recipeWoods(consumer, BVBlocks.Warped_Hyphae_Stairs.get(), BVBlocks.Warped_Hyphae_Slab.get(), BVBlocks.Warped_Hyphae_Wall.get(), BVBlocks.Warped_Hyphae_Fence.get(), BVBlocks.Warped_Hyphae_FenceGate.get(), Blocks.WARPED_HYPHAE, Blocks.WARPED_STEM);
+        recipeWoods(consumer, BVBlocks.Stripped_Oak_Wood_Stairs.get(), BVBlocks.Stripped_Oak_Wood_Slab.get(), BVBlocks.Stripped_Oak_Wood_Wall.get(), BVBlocks.Stripped_Oak_Wood_Fence.get(), BVBlocks.Stripped_Oak_Wood_FenceGate.get(), Blocks.STRIPPED_OAK_WOOD, Blocks.STRIPPED_OAK_LOG);
+        recipeWoods(consumer, BVBlocks.Stripped_Spruce_Wood_Stairs.get(), BVBlocks.Stripped_Spruce_Wood_Slab.get(), BVBlocks.Stripped_Spruce_Wood_Wall.get(), BVBlocks.Stripped_Spruce_Wood_Fence.get(), BVBlocks.Stripped_Spruce_Wood_FenceGate.get(), Blocks.STRIPPED_SPRUCE_WOOD, Blocks.STRIPPED_SPRUCE_LOG);
+        recipeWoods(consumer, BVBlocks.Stripped_Birch_Wood_Stairs.get(), BVBlocks.Stripped_Birch_Wood_Slab.get(), BVBlocks.Stripped_Birch_Wood_Wall.get(), BVBlocks.Stripped_Birch_Wood_Fence.get(), BVBlocks.Stripped_Birch_Wood_FenceGate.get(), Blocks.STRIPPED_BIRCH_WOOD, Blocks.STRIPPED_BIRCH_LOG);
+        recipeWoods(consumer, BVBlocks.Stripped_Jungle_Wood_Stairs.get(), BVBlocks.Stripped_Jungle_Wood_Slab.get(), BVBlocks.Stripped_Jungle_Wood_Wall.get(), BVBlocks.Stripped_Jungle_Wood_Fence.get(), BVBlocks.Stripped_Jungle_Wood_FenceGate.get(), Blocks.STRIPPED_JUNGLE_WOOD, Blocks.STRIPPED_JUNGLE_LOG);
+        recipeWoods(consumer, BVBlocks.Stripped_Acacia_Wood_Stairs.get(), BVBlocks.Stripped_Acacia_Wood_Slab.get(), BVBlocks.Stripped_Acacia_Wood_Wall.get(), BVBlocks.Stripped_Acacia_Wood_Fence.get(), BVBlocks.Stripped_Acacia_Wood_FenceGate.get(), Blocks.STRIPPED_ACACIA_WOOD, Blocks.STRIPPED_ACACIA_LOG);
+        recipeWoods(consumer, BVBlocks.Stripped_Dark_Oak_Wood_Stairs.get(), BVBlocks.Stripped_Dark_Oak_Wood_Slab.get(), BVBlocks.Stripped_Dark_Oak_Wood_Wall.get(), BVBlocks.Stripped_Dark_Oak_Wood_Fence.get(), BVBlocks.Stripped_Dark_Oak_Wood_FenceGate.get(), Blocks.STRIPPED_DARK_OAK_WOOD, Blocks.STRIPPED_DARK_OAK_LOG);
+        recipeWoods(consumer, BVBlocks.Stripped_Mangrove_Wood_Stairs.get(), BVBlocks.Stripped_Mangrove_Wood_Slab.get(), BVBlocks.Stripped_Mangrove_Wood_Wall.get(), BVBlocks.Stripped_Mangrove_Wood_Fence.get(), BVBlocks.Stripped_Mangrove_Wood_FenceGate.get(), Blocks.STRIPPED_MANGROVE_WOOD, Blocks.STRIPPED_MANGROVE_LOG);
+        recipeWoods(consumer, BVBlocks.Stripped_Crimson_Hyphae_Stairs.get(), BVBlocks.Stripped_Crimson_Hyphae_Slab.get(), BVBlocks.Stripped_Crimson_Hyphae_Wall.get(), BVBlocks.Stripped_Crimson_Hyphae_Fence.get(), BVBlocks.Stripped_Crimson_Hyphae_FenceGate.get(), Blocks.STRIPPED_CRIMSON_HYPHAE, Blocks.STRIPPED_CRIMSON_STEM);
+        recipeWoods(consumer, BVBlocks.Stripped_Warped_Hyphae_Stairs.get(), BVBlocks.Stripped_Warped_Hyphae_Slab.get(), BVBlocks.Stripped_Warped_Hyphae_Wall.get(), BVBlocks.Stripped_Warped_Hyphae_Fence.get(), BVBlocks.Stripped_Warped_Hyphae_FenceGate.get(), Blocks.STRIPPED_WARPED_HYPHAE, Blocks.STRIPPED_WARPED_STEM);
 
         all(consumer, BVBlocks.Calcite_Stairs.get(), BVBlocks.Calcite_Slab.get(), BVBlocks.Calcite_Wall.get(), Blocks.CALCITE, true);
         all(consumer, BVBlocks.Smooth_Basalt_Stairs.get(), BVBlocks.Smooth_Basalt_Slab.get(), BVBlocks.Smooth_Basalt_Wall.get(), Blocks.SMOOTH_BASALT, true);
@@ -295,13 +292,17 @@ public class BVRecipeProvider extends ExtendedRecipeProvider {
     }
 
     protected void smelting(Consumer<FinishedRecipe> consumer, ItemLike result, ItemLike ingredient) {
-        furnance(consumer, result, ingredient, 0.1f, 200);
+        smelting(consumer, result, ingredient, 0.1f, 200);
     }
-    protected void recipeWoods(Consumer<FinishedRecipe> consumer, StairBlock stairs, SlabBlock slab, WallBlock wall, ItemLike material, ItemLike altMaterial){
+    protected void recipeWoods(Consumer<FinishedRecipe> consumer, StairBlock stairs, SlabBlock slab, WallBlock wall, FenceBlock fence, FenceGateBlock fenceGate, ItemLike material, ItemLike altMaterial){
         stairs(consumer, stairs, material, false);
         slab(consumer, slab, material, false);
         wall(consumer, wall, material, false);
         wall(consumer, wall, altMaterial, false);
+        fence(consumer, fence, material);
+        fence(consumer, fence, altMaterial);
+        fenceGate(consumer, fenceGate, material);
+        fenceGate(consumer, fenceGate, altMaterial);
     }
 
     protected static ResourceLocation recipeID(ItemLike item) {return new ResourceLocation(BlockVariants.ModID, getItemName(item));}